requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1
redis==5.2.0
Requests==2.32.3
rq==1.16.2
